requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-filter==23.3
psycopg2-binary==2.9.11
python-dotenv==1.2.1
